pyproject.toml
[project]
name = "mlops-temporal-poc"
version = "0.1.0"
requires-python = ">=3.10,<4.0"
readme = "README.md"
dependencies = [
    "temporalio>=1.8.0,<2",
    "watchdog>=6.0.0,<7",
    "pandas>=2.2.3,<3",
    "pyyaml>=6.0.2,<7",
    "scikit-learn>=1.7.1,<2",
    "requests>=2.32.3,<3",
    "streamlit>=1.48.1,<2",
    "psycopg2-binary>=2.9.0,<3",
    "python-dotenv>=1.0.0,<2",
    "cryptography>=41.0.0,<42",
]